Refactored more neovide setup

+++ install.sh
@@ -152,46 +152,52 @@ setBubblyConfig() {
       echo "User added. Setup will now continue with the remaining steps."
     fi
 
-    else
-      echo "Skipping bubbly installation..."
+  else
+    echo "Skipping bubbly installation..."
   fi
 
 }
 
 # Neovide config setup function
 setNeovideConfig() {
-  # Install nvim if it does not exist
-  echo "Checking for nvim installation..."
-  nvimPath=$(which nvim)
+  # Configure neovim on user request
+  read -p "Would you like to configure the neovim IDE now? [Y,n]: " -i Y nvimReply
+  if [ $nvimReply -eq "y" || $nvimReply -eq "Y" ]; then
+    # Install nvim if it does not exist
+    echo "Checking for nvim installation..."
+    nvimPath=$(which nvim)
 
-  # Checking to see if nvim path is registered
-  if [ $nvimPath 2> /dev/null -eq "" ]; then
-    echo "Nvim not detected, installing nvim..."
-    sudo apt install nvim
+    # Checking to see if nvim path is registered
+    if [ $nvimPath 2> /dev/null -eq "" ]; then
+      echo "Nvim not detected, installing nvim..."
+      sudo apt install nvim
+    fi
+
+    # Copy neovim files to ~/.config/nvim
+    nvimPath="$userHome/.config/nvim/"
+    mkdir -p $nvimPath
+    echo "Copying neovim files to $nvimPath..."
+    cp -r nvim/* $nvimPath
+
+    # Ask to install neovide
+    read -p "Would you like to install NEOVIDE now? [Y,n]: " -i Y neovideReply
+    if [ $neovideReply -eq "y" || $neovideReply -eq "Y" ]; then
+      # Install neovide
+      echo "Installing prerequisites..."
+      sudo apt install -y curl \
+          gnupg ca-certificates git \
+          gcc-multilib g++-multilib cmake libssl-dev pkg-config \
+          libfreetype6-dev libasound2-dev libexpat1-dev libxcb-composite0-dev \
+          libbz2-dev libsndio-dev freeglut3-dev libxmu-dev libxi-dev libfontconfig1-dev \
+          libxcursor-dev
+
+      # Use cargo to install neovide
+      cargo install --git https://github.com/neovide/neovide
+    fi
+
+  else
+    echo "Skipping neovide installation..."
   fi
-
-  # Copy neovim files to ~/.config/nvim
-  nvimPath="$userHome/.config/nvim/"
-  mkdir -p $nvimPath
-  echo "Copying neovim files to $nvimPath..."
-  cp -r nvim/* $nvimPath
-
-  # Ask to install neovide
-  read -p "Would you like to install NEOVIDE now? [Y,n]: " -i Y neovideReply
-  if [ $neovideReply -eq "y" || $neovideReply -eq "Y" ]; then
-    # Install neovide
-    echo "Installing prerequisites..."
-    sudo apt install -y curl \
-        gnupg ca-certificates git \
-        gcc-multilib g++-multilib cmake libssl-dev pkg-config \
-        libfreetype6-dev libasound2-dev libexpat1-dev libxcb-composite0-dev \
-        libbz2-dev libsndio-dev freeglut3-dev libxmu-dev libxi-dev libfontconfig1-dev \
-        libxcursor-dev
-
-    # Use cargo to install neovide
-    cargo install --git https://github.com/neovide/neovide
-  fi
-
 }
 
 # DWM config setup function
@@ -269,13 +275,6 @@ setNeovideConfig
 # Do dwm config task
 setDWMConfig
 
-# Configure neovim on user request
-read -p "Would you like to configure the neovim IDE now? [Y,n]: " -i Y nvimReply
-if [ $nvimReply -eq "y" || $nvimReply -eq "Y" ]; then
-  configNvim &
-else
-  echo "Skipping neovide installation..."
-fi
 
 echo "Last step!"
 # Install dwm
